Functional!
Assuming correct file structure, and correct username/password inputs, the ledfromemail.py file is fully functional.

It assumes that there is a folder 'displays' within the working folder.  To start, one would require at least one file in the displays folder and/or coordinates(structured in the format of the ledmousetracer.html) in your email to read.